choose_disk: Ordering LVM vs Ceph-volume
Some ceph disks could be under LVM, so we first have to check for ceph
based LVMs.

This is achieved by calling disk_label() on the block device first and
then for each partition.

If no ceph data got found, then we can check for system based LVMs.

This patch is also adding a little of comments to better explain the
code.

Signed-off-by: Erwan Velu <erwan@redhat.com>